requirements.txt
Flask
google-generativeai
PyMuPDF
python-docx
matplotlib
numpy
pandas
scikit-learn
nltk
spacy
